SiTRAV_Srv/sitrav_backend/gestor_conductores/serializers.py
from rest_framework import serializers
from .models import Conductores

class ConductoresSerializer(serializers.ModelSerializer):
    class Meta:
        model = Conductores
        fields = '__all__'